refactor: rename tiles for bookmarks to BookmarkTile + restore small tile as general component
